% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_items.R
\name{add_archive_url}
\alias{add_archive_url}
\title{Add archive URLs from "Robust Link" attachments}
\usage{
add_archive_url(items)
}
\arguments{
\item{items}{A list or data frame of Zotero library items as returned by
\code{\link[=collection_items]{collection_items()}}. Must include all of the fields/columns \code{parentItem},
\code{itemType} and \code{title}.}
}
\value{
A \link[tibble:tbl_df-class]{tibble}.
}
\description{
Adds a top-level column \code{archiveUrl} with archive links extracted from item
attachments created by the \href{https://robustlinks.mementoweb.org/zotero/}{Zotero Robust Links} add-on.
}
\details{
An example of a "Robust Link" attachment can be viewed via \href{https://www.zotero.org/groups/4603023/zoterro_testing/items/KC6WFG78/attachment/EE7ZQ6PD/library}{this link}.
}
\examples{
library(magrittr)

# fetch some demo entry
collection_items(key = "AZBJE3R8",
                 user = zotero_group_id(4603023)) \%>\%
  # add archive link
  add_archive_url() \%>\%
  # drop "Robust Link" attachment
  dplyr::filter(!(itemType == "attachment" & title == "Robust Link")) \%>\%
  # show columns
  dplyr::select(key, url, archiveUrl)
}
\seealso{
Other Zotero items functions:
\code{\link{as_item_tibble}()},
\code{\link{collection_items}()}
}
\concept{items}
