% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_items.R
\name{collection_items}
\alias{collection_items}
\title{Fetch all items in a collection}
\usage{
collection_items(key, ...)
}
\arguments{
\item{key}{collection key, see \code{\link[=collections]{collections()}}}

\item{...}{other arguments passed to \code{\link[=zotero_api]{zotero_api()}}}
}
\description{
Fetch all items in a collection
}
