% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Fetch collection hierarchy}
\usage{
collections(...)
}
\arguments{
\item{...}{Further arguments passed on to \code{\link[=zotero_api]{zotero_api()}}}
}
\value{
A \link[tibble:tbl_df-class]{tibble} with columns:
\itemize{
\item \code{key}: Collection key.
\item \code{name}: Collection name.
\item \code{parent_key}: Key of the parent collection (\code{NA} if no parent collection exists).
}
}
\description{
Fetch all collections of a Zotero user or group library with their keys and
keys of parent collections (if any).
}
\examples{
collections(user = zotero_group_id(id = 197065))
}
