% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{items}
\alias{items}
\title{Fetch Zotero library items}
\usage{
items(collection_key = NULL, incl_children = TRUE, as_tibble = TRUE, ...)
}
\arguments{
\item{collection_key}{\href{https://www.zotero.org/support/dev/web_api/v3/basics#collections}{Collection key},
see \code{\link[=collections]{collections()}}. If \code{NULL}, all items in the library are returned.}

\item{incl_children}{Whether or not to include child items from
sub-collections beneath the \code{collection_key}'s or the library's (if
\code{collection_key = NULL}) top level.}

\item{as_tibble}{Whether or not to convert the resulting list to a tibble
(with nested list columns). If \code{FALSE}, a bare list is returned instead
which also retains all additional API metadata fields besides the actual
items data. See \code{\link[=as_item_tibble]{as_item_tibble()}} for details.}

\item{...}{Other arguments passed to \code{\link[=zotero_api]{zotero_api()}}}
}
\value{
A \link[tibble:tbl_df-class]{tibble} if \code{as_tibble = TRUE}, otherwise a list.

In both cases the returned object
\itemize{
\item has a \code{version} \link[=attr]{attribute} set to the \href{https://www.zotero.org/support/dev/web_api/v3/syncing}{Zotero library version number} returned by
the API.
\item is of length \code{0} if the Zotero library's content hasn't changed since the
version number specified in \code{modified_since}.
}

See section \emph{Caching} in \code{\link[=zotero_api]{zotero_api()}} for further details.
}
\description{
Retrieves items belonging to a Zotero user or group library, optionally
limited to a specific collection. Trashed items are never included.
}
\examples{
# items from public libraries can be fetched without an API key
items(collection_key = "MSNU4A4F", user = zotero_group_id(id = 197065))

# when no collection key is supplied, the whole library is fetched
items(user = zotero_group_id(id = 197065))

# optionally, the retrieval ca be limited to top-level items only
items(incl_children = FALSE, user = zotero_group_id(id = 197065))
}
\seealso{
Other Zotero items functions:
\code{\link{add_archive_url}()},
\code{\link{as_item_tibble}()},
\code{\link{write_bib}()}
}
\concept{items}
