% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{parse_results}
\alias{parse_results}
\title{Given a list of GET responses return a useful object}
\usage{
parse_results(x, ...)
}
\arguments{
\item{r}{}
}
\description{
Given a list of GET responses return a useful object
}
\keyword{internal}
