% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_credentials.R
\name{print.zotero_id}
\alias{print.zotero_id}
\title{Print Zotero user/group ID}
\usage{
\method{print}{zotero_id}(x, ...)
}
\arguments{
\item{x}{Zotero user/group ID}

\item{...}{other arguments}
}
\description{
Print Zotero user/group ID
}
\keyword{internal}
