% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_items.R
\name{save_collection}
\alias{save_collection}
\title{Export all items in a collection to file}
\usage{
save_collection(key, path, format = "bibtex", ...)
}
\arguments{
\item{key}{collection key, see \code{\link[=collections]{collections()}}}

\item{path}{Path to file.}

\item{format}{Bibliographic data format used for export. One of:
\itemize{
\item \code{"bibtex"},
\item \code{"biblatex"},
\item \code{"csljson"},
\item \code{"mods"},
\item \code{"refer"},
\item \code{"rdf_bibliontology"},
\item \code{"rdf_dc"},
\item \code{"rdf_zotero"},
\item \code{"ris"},
\item \code{"wikipedia"}
}

For details, see the \href{https://www.zotero.org/support/dev/web_api/v3/basics#export_formats}{relevant Zotero documentation}.}

\item{...}{other arguments passed to \code{\link[=zotero_api]{zotero_api()}}}
}
\description{
Writes a file of the chosen \code{format} to \code{path} with all items in a collection.
}
