% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{zotero_response_links}
\alias{zotero_response_links}
\title{Extract the links to subsequent queries (pages) from Zotero response}
\usage{
zotero_response_links(r, ...)
}
\arguments{
\item{r}{response

Returns named character vector with names from: first, prev, next, last. URL
named alternate leads to corresponding webpage.}
}
\description{
Extract the links to subsequent queries (pages) from Zotero response
}
\keyword{internal}
