% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoterro-package.R
\docType{package}
\name{zoterro-package}
\alias{zoterro}
\alias{zoterro-package}
\title{zoterro: Zotero Web API Interface}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Accessing Zotero web API to fetch/put/manipulate bibliographical data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mbojan/zoterro}
  \item Report bugs at \url{https://github.com/mbojan/zoterro/issues}
}

}
\author{
\strong{Maintainer}: Michal Bojanowski \email{michal2992@gmail.com} (\href{https://orcid.org/0000-0001-7503-852X}{ORCID})

Other contributors:
\itemize{
  \item Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID}) [contributor]
}

}
\keyword{internal}
