% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_items.R
\name{collection_items}
\alias{collection_items}
\alias{save_collection}
\title{Fetch all items of a collection}
\usage{
collection_items(key, format = "bibtex", ...)

save_collection(key, path, ...)
}
\arguments{
\item{key}{collection key, see \code{\link[=collections]{collections()}}}

\item{format}{output format}

\item{path}{path to bibtex file}
}
\description{
\itemize{
\item \code{save_collection()} - Write BibTeX file with all items in a collection
}
}
