% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_items.R
\name{collection_items}
\alias{collection_items}
\alias{save_collection}
\title{Fetch all items in a collection}
\usage{
collection_items(key, ...)

save_collection(key, path, ...)
}
\arguments{
\item{key}{collection key, see \code{\link[=collections]{collections()}}}

\item{...}{other arguments passed to \code{\link[=zotero_api]{zotero_api()}}}

\item{path}{path to file}
}
\description{
\itemize{
\item \code{save_collection()} - Write BibTeX file \code{path} with all items in a collection
}
}
