% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\title{Fetch collection hierarchy}
\usage{
collections(...)
}
\arguments{
\item{...}{passed to \code{\link[=zotero_api]{zotero_api()}}}
}
\value{
Data frame with columns:
\itemize{
\item \code{key} - collection key
\item \code{name} - collection name
\item \code{parent} - key of the parent collection or \code{NA}
}
}
\description{
Fetch all collections of a Zotero user with their keys and keys of parent
collections (if any).
}
