% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{write_bib}
\alias{write_bib}
\title{Export Zotero library items to bibliography file}
\usage{
write_bib(
  collection_key = NULL,
  incl_children = TRUE,
  path,
  format = "bibtex",
  modified_since = NULL,
  ...
)
}
\arguments{
\item{collection_key}{\href{https://www.zotero.org/support/dev/web_api/v3/basics#collections}{Collection key},
see \code{\link[=collections]{collections()}}. If \code{NULL}, all items in the library are returned.}

\item{incl_children}{Whether or not to include child items from
sub-collections beneath the \code{collection_key}'s or the library's (if
\code{collection_key = NULL}) top level.}

\item{path}{Path to file.}

\item{format}{Bibliographic data format used for export. One of:
\itemize{
\item \code{"bibtex"},
\item \code{"biblatex"},
\item \code{"csljson"},
\item \code{"mods"},
\item \code{"refer"},
\item \code{"rdf_bibliontology"},
\item \code{"rdf_dc"},
\item \code{"rdf_zotero"},
\item \code{"ris"},
\item \code{"wikipedia"}
}

For details, see the \href{https://www.zotero.org/support/dev/web_api/v3/basics#export_formats}{relevant Zotero documentation}.}

\item{modified_since}{Optional \href{https://www.zotero.org/support/dev/web_api/v3/syncing}{Zotero library version number}. If the
Zotero library's content hasn't changed since the specified version
number, nothing will be written to \code{path} (and the performed API request
will be significantly faster). See section \emph{Caching} in \code{\link[=zotero_api]{zotero_api()}} for
further details.}

\item{...}{Other arguments passed to \code{\link[=zotero_api]{zotero_api()}}}
}
\value{
A character scalar of the Zotero Library items in the specified \code{format},
invisibly.

Note that it is of length \code{0} if the Zotero library's content
hasn't changed since the version number specified in \code{modified_since}.
Otherwise it has a \code{version} \link[=attr]{attribute} set to the \href{https://www.zotero.org/support/dev/web_api/v3/syncing}{Zotero library version number} returned
by the API.

See section \emph{Caching} in \code{\link[=zotero_api]{zotero_api()}} for further details.
}
\description{
Writes a file of the chosen bibliography \code{format} to \code{path} with items
belonging to a Zotero user or group library, optionally limited to a
specific collection.
}
\seealso{
Other Zotero items functions:
\code{\link{add_archive_url}()},
\code{\link{as_item_tibble}()},
\code{\link{items}()}
}
\concept{items}
