% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{zotero_api}
\alias{zotero_api}
\title{Make Zotero API request}
\usage{
zotero_api(
  base_url = "https://api.zotero.org",
  query = NULL,
  fetch_subsequent = TRUE,
  path = NULL,
  user = zotero_usr(),
  api_key = zotero_key(),
  verbose = getOption("zoterro.verbose", FALSE),
  ...
)
}
\arguments{
\item{base_url}{Base URL of the \href{https://www.zotero.org/support/dev/web_api/v3/basics#base_url}{Zotero Web API}.}

\item{query}{A named list of \href{https://www.zotero.org/support/dev/web_api/v3/basics#general_parameters}{Zotero Web API URL parameters},
passed on to \code{\link[httr:modify_url]{httr::modify_url()}}.

Note that the maximum \code{limit} is \code{100} results. Set
\code{fetch_subsequent = TRUE} (default) to retrieve all results.}

\item{fetch_subsequent}{Whether or not to also retrieve results beyond
\code{query$start + query$limit}.}

\item{path}{\href{https://www.zotero.org/support/dev/web_api/v3/basics#resources}{Zotero Web API resource path}
without the \verb{<userOrGroupPrefix>/}. e.g. \code{items/top}.}

\item{user}{Object returned by \code{\link[=zotero_user_id]{zotero_user_id()}} or \code{\link[=zotero_group_id]{zotero_group_id()}}.}

\item{api_key}{Personal \href{https://www.zotero.org/support/dev/web_api/v3/basics#authentication}{Zotero Web API key}.
See \code{\link[=zotero_key]{zotero_key()}}. If \code{NULL}, authentication is omitted, resulting in
limited access to public Zotero libraries only.}

\item{verbose}{Whether or not to print called API URLs to the console.}

\item{...}{Further arguments passed on to \code{\link[httr:RETRY]{httr::RETRY()}}.}
}
\value{
List of \code{response} objects (c.f. \code{\link[httr:GET]{httr::GET()}}).
}
\description{
Fetch data from Zotero using API ver. 3. If the result is broken into
multiple parts, multiple requests are made to fetch everything.
}
\details{
The \code{user} argument expects a Zotero user or group ID. Use \code{\link[=zotero_user_id]{zotero_user_id()}}
or \code{\link[=zotero_group_id]{zotero_group_id()}} to pass it. By default \code{\link[=zotero_usr]{zotero_usr()}} is called which
fetches the ID from the \R option \code{zoterro.user} or the environment variable
\code{ZOTERO_USER}.

The URL of the request will contain the appropriate user/group ID prefix
which will be combined with \code{path} or \code{query} when supplied.

By default, \code{zotero_api()} respects the following \R options:
\itemize{
\item \strong{\code{zoterro.user}} (defaults to \code{\link[=zotero_usr]{zotero_usr()}}): Zotero user ID.
\item \strong{\code{zoterro.key}} (defaults to \code{\link[=zotero_key]{zotero_key()}}): Personal \href{https://www.zotero.org/support/dev/web_api/v3/basics#authentication}{Zotero Web API key}.
\item \strong{\code{zoterro.verbose}} (defaults to \code{FALSE}): Whether or not to print
called API URLs to the console.
}
}
\examples{
\dontrun{
# Fetch groups for the default user
zotero_api(path = "groups")

# Fetch top-level collections for the group with ID=12345
zotero_api(path = "collections/top", user = zotero_group_id(12345))
}
}
