% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{zotero_api}
\alias{zotero_api}
\title{Make Zotero API request}
\usage{
zotero_api(base_url = "https://api.zotero.org", ...)
}
\arguments{
\item{base_url}{API URL}

\item{...}{For \code{\link[=zotero_api]{zotero_api()}} passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
List of class \code{zotero_api} with elements \code{content} and \code{response}.
}
\description{
Make Zotero APIv3 request.
}
