% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{zotero_api}
\alias{zotero_api}
\title{Make Zotero API request}
\usage{
zotero_api(
  base_url = "https://api.zotero.org",
  query = NULL,
  path = NULL,
  user = zotero_usr(),
  ...
)
}
\arguments{
\item{base_url}{API URL}

\item{path, query}{passed to \code{\link[=modify_url]{modify_url()}}}

\item{user}{object returned by \code{\link[=zotero_user_id]{zotero_user_id()}} or \code{\link[=zotero_group_id]{zotero_group_id()}}}

\item{...}{For \code{\link[=zotero_api]{zotero_api()}} passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
List of \code{response} objects (c.f. \code{\link[httr:GET]{httr::GET()}}).
}
\description{
Make Zotero APIv3 request.
}
\details{
Functions are responsive to the following options:
\itemize{
\item \code{zoterro.verbose} - (default \code{FALSE}) give more feedback when running
\item \code{zoterro.sleep} - (default 1) sleep time between requests, see \code{\link[=Sys.sleep]{Sys.sleep()}}
}
}
