% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{zotero_api}
\alias{zotero_api}
\title{Make Zotero API request}
\usage{
zotero_api(
  base_url = "https://api.zotero.org",
  query = NULL,
  path = NULL,
  user = zotero_usr(),
  ...
)
}
\arguments{
\item{base_url}{API URL}

\item{path, query}{passed to \code{\link[=modify_url]{modify_url()}}}

\item{user}{object returned by \code{\link[=zotero_user_id]{zotero_user_id()}} or \code{\link[=zotero_group_id]{zotero_group_id()}}}

\item{...}{For \code{\link[=zotero_api]{zotero_api()}} passed to \code{\link[httr:GET]{httr::GET()}}.}
}
\value{
List of \code{response} objects (c.f. \code{\link[httr:GET]{httr::GET()}}).
}
\description{
Fetch data from Zotero using API ver. 3. If the result is broken into
multiple parts, multiple requests are made to fetch everything.
}
\details{
The \code{user} argument expects Zotero user or group ID. Use \code{\link[=zotero_user_id]{zotero_user_id()}}
or \code{\link[=zotero_group_id]{zotero_group_id()}} to pass it. By default \code{\link[=zoter_usr]{zoter_usr()}} is called which
fetches the ID from the option or the environment variable.

The URL of the request will contain the appropriate user/group ID prefix
which will be combined with \code{path} or \code{query} when supplied.

The function is responsive to the following options:
\itemize{
\item \code{zoterro.verbose} - (default \code{FALSE}) give more feedback when running
\item \code{zoterro.sleep} - (default 1) sleep time between requests, see \code{\link[=Sys.sleep]{Sys.sleep()}}
}
}
