% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_credentials.R
\name{zotero_credentials}
\alias{zotero_credentials}
\alias{zotero_usr}
\alias{zotero_key}
\title{Acess Zotero username and key from environment variables.}
\usage{
zotero_usr()

zotero_key()
}
\description{
\itemize{
\item \code{zotero_usr()} - return Zotero user ID
}

\itemize{
\item \code{zotero_key()} - return Zotero key
}
}
\details{
Best practice is to store both the user ID and the key in
read-protected \code{.Renviron} as environment variables. The two functions
\code{ZOTERO_USER} and \code{ZOTERO_KEY} respectively.
}
