% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_credentials.R
\name{zotero_credentials}
\alias{zotero_credentials}
\alias{zotero_usr}
\alias{zotero_key}
\title{Get Zotero username and key}
\usage{
zotero_usr()

zotero_key()
}
\description{
\itemize{
\item \code{zotero_usr()} - return Zotero user ID
}

\itemize{
\item \code{zotero_key()} - return Zotero key
}
}
\details{
These functions look for Zotero user ID and key by looking into the
following places and returning as soon as a value found:
\itemize{
\item Options \code{zotero.user} and \code{zotero.key} respectively
\item Environment variables \code{ZOTERO_USER} and \code{ZOTERO_KEY} respectively.
}

Best practice is to store both the user ID and the key in read-protected
\verb{~/.Renviron} as environment variables.
}
